[IMP] web: add FloatTime component
Allow FloatTime to be used as a standalone owl component outside of
field context.

task-5072614
