+++ addons/hr_holidays/static/src/components/float_time_selection/float_time_selection.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-    <t t-name="hr_holidays.FloatTimeSelectionField" t-inherit="web.FloatTimeField">
+    <t t-name="hr_holidays.FloatTimeSelectionField" t-inherit="web.FloatTime">
         <xpath expr="//input" position="attributes">
             <attribute name="t-on-input">this.handleInputChange</attribute>
             <attribute name="t-on-click">this.onCharHoursClick</attribute>
