+++ addons/web/static/src/views/fields/float_time/float_time_field.js
@@ -1,98 +1,24 @@
-import { useState } from "@web/owl2/utils";
 import { _t } from "@web/core/l10n/translation";
 import { registry } from "@web/core/registry";
-import { formatFloatTime } from "../formatters";
 import { useInputField } from "../input_field_hook";
 import { standardFieldProps } from "../standard_field_props";
-import { useNumpadDecimal } from "../numpad_decimal_hook";
 import { parseFloatTime } from "../parsers";
+import { FloatTime } from "./float_time";
 
-import { Component } from "@odoo/owl";
-import { usePopover } from "@web/core/popover/popover_hook";
+export class FloatTimeField extends FloatTime {
+    static props = { ...standardFieldProps, ...FloatTime.props };
 
-export class FloatTimeField extends Component {
-    static template = "web.FloatTimeField";
-    static props = {
-        ...standardFieldProps,
-        showSeconds: { type: Boolean, optional: true },
-        numeric: { type: Boolean, optional: true },
-        unit: {
-            type: [{ value: "hours" }, { value: "minutes" }, { value: "seconds" }],
-            optional: true,
-        },
-    };
-    static defaultProps = {
-        numeric: false,
-        unit: "hours",
-    };
-
-    setup() {
-        this.inputFloatTimeRef = useInputField({
+    getUseInputRef() {
+        return useInputField({
             getValue: () => this.formattedValue,
             refName: "numpadDecimal",
             parse: (v) => parseFloatTime(v, this.props.unit),
         });
-
-        this.state = useState({
-            formattedResult: "",
-        });
-        this.resultPopover = usePopover(DurationPopover, {
-            position: "bottom",
-        });
-        useNumpadDecimal();
     }
 
-    onValueChange(ev) {
-        const currentInput = ev.target.value;
-        this.state.formattedResult = formatFloatTime(
-            parseFloatTime(currentInput, this.props.unit),
-            {
-                showSeconds: this.props.showSeconds,
-                numeric: this.props.numeric,
-                unit: this.props.unit,
-            }
-        );
-        if (currentInput === this.state.formattedResult && this.resultPopover.isOpen) {
-            this.resultPopover.close();
-        } else if (currentInput !== this.state.formattedResult && !this.resultPopover.isOpen) {
-            this.resultPopover.open(this.inputFloatTimeRef.el, {
-                state: this.state,
-            });
-        }
+    get floatValue() {
+        return this.props.record.data[this.props.name];
     }
-
-    openPopover() {
-        const duration = parseFloatTime(this.inputFloatTimeRef.el.value, this.props.unit);
-        this.state.formattedResult = formatFloatTime(duration, {
-            showSeconds: this.props.showSeconds,
-            numeric: this.props.numeric,
-            unit: this.props.unit,
-        });
-        this.resultPopover.open(this.inputFloatTimeRef.el, {
-            state: this.state,
-        });
-    }
-
-    closePopover() {
-        this.resultPopover.close();
-    }
-
-    get formattedValue() {
-        const value = this.props.record.data[this.props.name];
-        return formatFloatTime(value, {
-            showSeconds: this.props.showSeconds,
-            numeric: this.props.numeric,
-            unit: this.props.unit,
-        });
-    }
-}
-
-class DurationPopover extends Component {
-    static template = "web.DurationPopover";
-    static props = {
-        state: { type: Object, optional: true },
-        close: { type: Function, optional: true },
-    };
 }
 
 export const floatTimeField = {
